a04-move-repo-traits-to-per-file: `src/repository/mod.rs` is 321 lines o
