+++ openspec/changes/archive/2026-05-16-a04-move-repo-traits-to-per-file/design.md
